[diffusion] refactor: split component_loader into component-wise files
upd

rebase
