+++ python/sglang/multimodal_gen/runtime/pipelines_core/stages/decoding.py
@@ -10,7 +10,7 @@ import weakref
 import torch
 
 from sglang.multimodal_gen.runtime.distributed import get_local_torch_device
-from sglang.multimodal_gen.runtime.loader.component_loader import VAELoader
+from sglang.multimodal_gen.runtime.loader.vae_loader import VAELoader
 from sglang.multimodal_gen.runtime.models.vaes.common import ParallelTiledVAE
 from sglang.multimodal_gen.runtime.pipelines_core.schedule_batch import OutputBatch, Req
 from sglang.multimodal_gen.runtime.pipelines_core.stages.base import (
